added safguard to make sure that there is no pre-existing folder for the task to be run, unless the -F|--FORCE mode is toggled on.

+++ scripts/pipeline/pantagruel_pipeline_01_homologous_seq_families.sh
@@ -13,12 +13,27 @@ if [ -z "$1" ] ; then echo "missing mandatory parameter: pantagruel config file"
 envsourcescript="$1"
 source ${envsourcescript}
 
+checkfoldersafe (){
+  if [ -d ${1} ] ; then
+    if [ ${runmode} != 'force' ] ; then
+      echo "Task folder ${1} already exists; will stop here rther then overwritting data."
+      echo "If you want previous data to be reased, use FORCE mode with -F|--FORCE option."
+    else
+      echo "Task folder ${1} already exists; FORCE mode is on: WILL ERASE the folder and write new result in its place"
+      rm -r ${1}
+    fi
+  else
+    mkdir ${1}
+  fi
+}
+
+checkfoldersafe ${seqdb}
+
 #############################
 ## 01. Homologous Sequence db
 #############################
 
 ## extract all the protein sequences into single proteome fasta files
-mkdir -p ${seqdb}
 
 rm -f ${allfaarad}*
 for ass in `ls ${assemblies}` ; do
